q2/requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
statsmodels>=0.14.0
pyyaml>=6.0
holidays>=0.35
prophet>=1.1.4
tensorflow>=2.13.0
scipy>=1.10.0

# Amazon Chronos - install from GitHub
# Run separately: pip install git+https://github.com/amazon-science/chronos-forecasting.git